Enable KV cache reuse across conversations (#51)
Currently, when user enter an input, the behavior is to clear KV cache
and collect all previous conversation history as input prompt to the
model. It becomes extremely slow when the history is large. But clearly,
there exists some redundant kv computation for previous conversations.
We could utilize the kv cache from the last round of conversation to
avoid kv recomputation. This PR enables such optimization.