Update Chrome extension to work with both WebGPU and local server (#182)
This PR updates the Chrome extension to run in two modes:
- Using the `ChatRestModule`, which uses a local/hosted server to run
the model.
- This requires the user to launch the server before using the Chrome
extension, which interacts with the server using the REST API.
- The `ChatRestModule` will be created in the background service worker,
so it should persist between interactions.
- Using the `ChatModule`, which runs the model using WebGPU.
- This does not require any additional user actions before using the
extension.
- The `ChatModule` cannot be created in the background service worker,
so it needs to be loaded from cache every time the user interacts with
the extension. This takes several seconds.

Some additional observations / limitations:
- As far as I can tell, Chrome extensions can only
[support](https://developer.chrome.com/docs/extensions/mv3/service_workers/)
background service workers, which don't currently have support for
WebGPU yet, as mentioned in
https://github.com/gpuweb/gpuweb/issues/4197. This means that the WebGPU
mode will only work from the main script and not the background worker.
- This approach does not work with Manifest v3 because of the following
error:
```
Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self'".
```
I suspect it is because we are evaluating some third party code which is
not in the extension. There seems to be a dependency on 'unsafe-eval' in
the Content Security Policy, but according to the Chrome extension
[documentation](https://developer.chrome.com/docs/extensions/mv3/manifest/content_security_policy/),
this is not allowed starting v3. I managed to get it to work in Manifest
v2, but Manifest v2 will be getting deprecated this year.