[Wasm] Add wasm versioning, expose all prebuilts, add 0.2.30 wasms (#355)
As part of the effort to create a better user experience in the upcoming
new release of WebLLM, we add wasm versioning and other changes in this
PR:
- Adds wasm versioning (see discussion below)
- TLDR: users are no longer forced to update WebLLM npm whenever we add
breaking changes to prebuilt models
- Expose all prebuilt wasms as `webllm.prebuiltAppConfig`
- Previously `prebuiltAppConfig` only has three models for demonstration
purposes, and we store all prebuilt model records under
`examples/simple-chat`. We now move the full prebuilt list to
`prebuiltAppConfig`.
- Rename `local_id` to `model_id` for clearer semantics and to match
other platforms of mlc-llm (iOS and Android)
- Release 0.2.30 wasms
  - Compiled with
- MLC-LLM at
https://github.com/mlc-ai/mlc-llm/commit/63fc9723c17261ad839f3f8267fdc7c89707ebb6
- TVM at
https://github.com/apache/tvm/commit/b4b97f8278c614dd2602c4f627d1229a015e00cb
- i.e. Relax at
https://github.com/mlc-ai/relax/commit/4bdcf425f762080614a218b229b3e4c38828bca0
- Main changes include the support of grammar, all use PagedKVCache, and
various other TVM runtime functions
- Also see https://github.com/mlc-ai/binary-mlc-llm-libs/pull/110

### Motivation of WASM Versioning
We introduce WebLLM's wasm versioning (we will use "wasm" and "model
library" interchangeably). That is, we preserve model libraries that are
compatible with previous npm versions instead of overriding them.

The motivation is that, sometimes if we update WebLLM's npm version, we
update model libraries as well. However, if we do not preserve the old
wasms, the new WebLLM npm might not be compatible with the old wasms. It
would be very tedious and hard to make sure a new npm version is
compatible with all previous wasms. We also do not want to force
WebLLM's users to update the npm whenever a new one is released.

However, not all new npm version needs to update the model libraries.
Therefore, it is only updated when needed, i.e. it is possible that
0.2.31 WebLLM npm uses 0.2.30 wasms. We refer to `prebuiltAppConfig`
exported by each WebLLM's npm for the only source of truth:
https://github.com/mlc-ai/web-llm/blob/main/src/config.ts -- what wasms
this npm version is using.

### Compiling your own model library (i.e. wasm)
For users who want to compile their own libraries for a specific WebLLM
npm version, if they compile at the same commits for MLC-LLM and TVM
stated above (we will mention the commits for all future wasm release),
they should expect the same behavior / compatibility with the provided
libraries.