[Grammar] Support response_format json mode in chatCompletion (#350)
We support json formatting in chatCompletion, essentially the option
`response_format: { type: "json_object" }` in a `chatCompletionRequest`.
By adding this field, it is guaranteed that the model would generate in
json format.

We also observe no performance degradation, possibly due to how
bitmasking is performed with WebGPU kernels.

This requires wasms to be built with MLC-LLM after
https://github.com/mlc-ai/mlc-llm/pull/2061.

From this PR on, old wasms would not be compatible with the newer
version of WebLLM npms. We will add wasm versioning soon so that old
wasms can work with old web-llm npm.