[Config] Enhance ModelRecord (#435)
There are three changes to `ModelRecord` this PR brings:

### 1. Update model ids to match HF repo name
We rename `modelId` in `webllm.prebuiltAppConfig` to be the exact same
as the HF repo name. For most models, that means we simply append `-MLC`
to the `modelId`. For the low-context version of the model, we would
have `{HF-repo}-1k`, suggesting 1k context length.

As a result, we rename Phi2 and phi1.5 models since their `modelId` did
not match with the repo name
- `Phi2-q4f32_1` → `phi-2-q4f32_1-MLC`
- `Phi1.5-q4f16_1` → `phi-1_5-q4f16_1-MLC`

### 2. Rename `model_url` and `model_lib_url` to `model` and `model_lib`
To better match with other platforms of MLC-LLM (e.g. iOS, Android), we
rename the `ModelRecord` fields.

### 3. Remove `resolve/main` from `model` URL
Instead of
`"https://huggingface.co/mlc-ai/Llama-3-8B-Instruct-q4f16_1-MLC/resolve/main/"`,
we now make it
`"https://huggingface.co/mlc-ai/Llama-3-8B-Instruct-q4f16_1-MLC/"`; note
the trailing `/` will be appended by us if it is not there.

### Example
As an example, we would have:
```typescript
    {
      model: "https://huggingface.co/mlc-ai/Llama-3-8B-Instruct-q4f16_1-MLC",
      model_id: "Llama-3-8B-Instruct-q4f16_1-MLC",
      model_lib: "path/to/Llama-3-8B-Instruct-q4f16_1-ctx1k_cs1k-webgpu.wasm",
    },
```
instead of 
```typescript
    {
      model_url: "https://huggingface.co/mlc-ai/Llama-3-8B-Instruct-q4f16_1-MLC/resolve/main/",
      model_id: "Llama-3-8B-Instruct-q4f16_1",
      model_lib_url: "path/to/Llama-3-8B-Instruct-q4f16_1-ctx4k_cs1k-webgpu.wasm",
    },
```

---------

Co-authored-by: Nestor Qin <imba.qxy@gmail.com>