[Streaming] Replace runtimeStatsText() with usage in examples (#460)
Recently we added `usage` to the last chunk in streaming if the user
specifies `streamOptions: { include_usage: True}` in the request, to be
compatible with the latest OpenAI API; for more see
https://github.com/mlc-ai/web-llm/pull/456.

This PR updates our streaming examples to use `chunk.usage` instead of
`runtimeStatsText()`. It is expected that we will deprecate
`runtimeStatsText()` in the future. Currently, only low-level API
`forwardTokenAndSample()` still needs it as they do not use OpenAI API
(e.g. `examples/logit-processor`).

Such examples include:
- simple-chat-ts
- simple-chat-js
- next-simple-chat
- streaming (already updated)